syntax: pull up ExternalItemOwner into `crate::ast`
